Put ContextMenu for ChatList into own file, remove console.log, some refactoring
